Only upload file to device if remote is different
This PR adds the ability to check if the remote file is different to the local file on upload. If its the same it can skip uploading to save time. 
For a single file upload this will increase the time to upload it takes a bit under a second to check if they are different plus the time to upload. However, when working on larger projects with multiple files, if you are only modifying a couple at a time it can speed up the upload process as it will only upload the files that need to be uploaded. It also means you don't need to upload each file individually.

You have to add the `--different` option to the script call otherwise it will use the default to uploading all files. 

Just wanted everyone's opinion on whether they see themselves using this or just prefer the old way.

If it was to be implemented I would recommend the run configuration to have a checkbox on whether it is enabled or not. I would also recommend disabling it for single file uploads and save it for project level or other multi-file uploads.